Move find_xml to new xml_helper.py
